React FastAPI with Auth Mantine/server/pyproject.toml
[project]
name = "mantine-full-stack-with-auth-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "cryptography>=44.0.3",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
